requirements.txt
streamlit
openai==0.27.8
python-dotenv
posthog
Flask==3.1.0
sendgrid==6.11.0
gunicorn==21.2.0
stripe==8.9.0
qrcode
pillow
twilio